♻️ 커스텀 페이지 리팩토링
