+++ src/components/Photographers/Custom/Data/IconList.js
@@ -1,0 +1,29 @@
+import instagram from "../../../../assets/photograph/instagram.png";
+import twitter from "../../../../assets/photograph/x.png";
+import kakao from "../../../../assets/photograph/kakao.png";
+import blog from "../../../../assets/photograph/blog.png";
+import home from "../../../../assets/photograph/home.png";
+
+// sns 아이콘 리스트
+export const icon_src = [
+  {
+    src: instagram,
+    text: "@ instagram",
+  },
+  {
+    src: twitter,
+    text: "@ twitter",
+  },
+  {
+    src: kakao,
+    text: "@ kakao channel",
+  },
+  {
+    src: blog,
+    text: "@ naver blog",
+  },
+  {
+    src: home,
+    text: "@ homepage",
+  },
+];
